Move day-types to data-inserter root
The day type id's are hardcoded values that are populated with migrations, we use the same id's everywhere
and we do not want to expose the datasets so a more logical place would be in the data-inserter root.
