+++ test/hasura/hsl/timetablesdb/datasets/additional-sets/timetables/draftSunApril2024Dataset.ts
@@ -1,9 +1,7 @@
-import {
-  defaultDayTypeIds,
-  journeyPatternRefsByName,
-} from 'generic/timetablesdb/datasets/defaultSetup';
+import { journeyPatternRefsByName } from 'generic/timetablesdb/datasets/defaultSetup';
 import { TimetablePriority } from 'generic/timetablesdb/datasets/types';
 import { DateTime } from 'luxon';
+import { defaultDayTypeIds } from '../../../../../timetables-data-inserter/day-types';
 import { buildHslVehicleScheduleFrame } from '../../factories';
 import { HslTimetablesDbTables } from '../../schema';
