+++ test/hasura/hsl/timetablesdb/datasets/defaultSetup/default-timetables-dataset.ts
@@ -1,4 +1,3 @@
-import { defaultDayTypeIds } from 'generic/timetablesdb/datasets/defaultSetup';
 import { defaultTimetablesDataset as defaultGenericTimetablesDataset } from 'generic/timetablesdb/datasets/defaultSetup/default-timetables-dataset';
 import {
   TimetablePriority,
@@ -6,6 +5,7 @@ import {
 } from 'generic/timetablesdb/datasets/types';
 import { DateTime } from 'luxon';
 import { mergeTimetablesDatasets } from 'timetables-data-inserter';
+import { defaultDayTypeIds } from '../../../../timetables-data-inserter/day-types';
 import { DayOfWeek } from '../types';
 
 const defaultTimetablesDatasetHslAdditions = {
