+++ test/hasura/hsl/timetablesdb/integration-tests/function-get-timetable-versions-by-journey-pattern-ids/get-timetable-versions-by-journey-pattern-ids.spec.ts
@@ -2,7 +2,6 @@ import * as config from '@config';
 import * as db from '@util/db';
 import { closeDbConnection, createDbConnection, DbConnection } from '@util/db';
 import { setupDb } from '@util/setup';
-import { defaultDayTypeIds } from 'generic/timetablesdb/datasets/defaultSetup';
 import {
   draftSunApril2023Timetable,
   draftSunApril2024Timetable,
@@ -22,6 +21,7 @@ import { DateTime } from 'luxon';
 import {
   buildHslTimetablesDataset,
   createHslTableData,
+  defaultDayTypeIds,
   mergeTimetablesDatasets,
 } from 'timetables-data-inserter';
 import { createHslTimetablesDatasetHelper } from 'timetables-data-inserter/hsl/timetables-dataset-helper';
