+++ test/hasura/hsl/timetablesdb/integration-tests/function-get_timetables_and_substitute_operating_days/get-timetables-and-substitute-operating-days.spec.ts
@@ -2,10 +2,7 @@ import * as config from '@config';
 import * as db from '@util/db';
 import { closeDbConnection, createDbConnection, DbConnection } from '@util/db';
 import { setupDb } from '@util/setup';
-import {
-  defaultDayTypeIds,
-  journeyPatternRefsByName,
-} from 'generic/timetablesdb/datasets/defaultSetup';
+import { journeyPatternRefsByName } from 'generic/timetablesdb/datasets/defaultSetup';
 import {
   draftSunApril2024Timetable,
   expressBusServiceSubstitutesSaturday20230520Dataset,
@@ -25,6 +22,7 @@ import { DateTime } from 'luxon';
 import {
   buildHslTimetablesDataset,
   createHslTableData,
+  defaultDayTypeIds,
   mergeTimetablesDatasets,
 } from 'timetables-data-inserter';
 import { createHslTimetablesDatasetHelper } from 'timetables-data-inserter/hsl/timetables-dataset-helper';
