+++ test/hasura/timetables-data-inserter/day-types.ts
