+++ test/hasura/timetables-data-inserter/generic/models/vehicle-service.ts
@@ -1,9 +1,9 @@
-import { defaultDayTypeIds } from 'generic/timetablesdb/datasets/defaultSetup/day-types';
 import {
   VehicleScheduleFrame,
   VehicleService,
 } from 'generic/timetablesdb/datasets/types';
 import { omit } from 'lodash';
+import { defaultDayTypeIds } from 'timetables-data-inserter/day-types';
 import { TimetablesDatasetInput } from 'timetables-data-inserter/types';
 import { assignId } from 'timetables-data-inserter/utils';
 import { VehicleServiceInput, VehicleServiceOutput } from '../types';
