+++ test/hasura/timetables-data-inserter/index.ts
@@ -1,3 +1,4 @@
+export * from './day-types';
 export * from './db';
 export * from './generic';
 export * from './hsl';
